feat: be able to share the hauk url from the gotify notification
This will be nice to be able to share the url with chat apps for example without open the link ourself.

I did the gotify android PR to make this working.

This can be merge before the gotify android but will not work while not merged https://github.com/gotify/android/pull/236.